Refactor incentive module and add details page
This refactor improves the incentive module by adding a new
"getIncentiveDetailsPage" function to fetch and render incentive
details. It uses an aggregate method to retrieve the list of months and
employees from the incentive collection. If successful, the fetched data
is passed to the "details.ejs" template for rendering. In case of an
error, a 400 status with the error message is sent. This change enhances
the user experience by providing a dedicated page to view incentive
details.

Additionally, some CSS classes were added to "incentive.css" to style
the elements of the details page. The form elements in the total
calculation page were modified to use dropdowns instead of input fields,
and some new CSS classes were added to "shared.css" to style radio
buttons and enhance the appearance.

These changes were made in the following files:
- controllers/incentive.ts
- public/css/incentive.css
- public/css/shared.css
- public/js/incentive.js
- routes/Incentive.ts
- views/Incentive/index.ejs
- views/Incentive/total.ejs

No related issue references.